@@ -1,21 +1,25 @@
 """Agent pool manager for concurrent issue processing."""
 
 import asyncio
-from dataclasses import dataclass
-from datetime import datetime, timedelta
-from enum import Enum
-from typing import Any
-from pathlib import Path
+import json
 import time
+from dataclasses import dataclass
+from datetime import UTC, datetime, timedelta
+from enum import Enum
+from pathlib import Path
+from typing import Any
 
 import structlog
+from fastmcp import Client as McpClient
 
-from ace.config.settings import get_settings
+from ace.agents.manager_agent import ManagerAgent
 from ace.config.secrets import resolve_github_token
+from ace.config.settings import get_settings
 from ace.github.api_client import GitHubAPIClient
 from ace.github.issue_queue import Issue, IssueQueue
 from ace.github.projects_v2 import ProjectsV2Client
 from ace.github.status_manager import IssueStatus
+from ace.logging_utils import log_key_event
 from ace.metrics import metrics
 from ace.orchestration.graph import get_compiled_graph
 from ace.orchestration.state import WorkerState
@@ -25,11 +29,68 @@ from ace.workspaces.tmux_ops import (
     parse_issue_from_session,
     session_name_for_issue,
 )
-from fastmcp import Client as McpClient
 
 logger = structlog.get_logger(__name__)
 
 MAX_CONCURRENT_AGENTS = 5
+
+
+def _extract_mcp_items(resp: Any) -> list[dict[str, Any]]:
+    """Normalize MCP tool responses into a list of issue-like dicts."""
+    if resp is None:
+        return []
+
+    for attr in ("structured_content", "structuredContent"):
+        structured = getattr(resp, attr, None)
+        if isinstance(structured, dict) and isinstance(structured.get("result"), list):
+            return structured["result"]
+
+    content = getattr(resp, "content", None)
+    if isinstance(content, list):
+        for part in content:
+            if isinstance(part, dict) and part.get("type") == "text":
+                try:
+                    parsed = json.loads(part.get("text", "[]"))
+                except json.JSONDecodeError:
+                    continue
+                if isinstance(parsed, list):
+                    return parsed
+    if isinstance(content, str):
+        try:
+            parsed = json.loads(content)
+        except json.JSONDecodeError:
+            parsed = None
+        if isinstance(parsed, list):
+            return parsed
+
+    if isinstance(resp, dict):
+        if isinstance(resp.get("result"), list):
+            return resp["result"]
+        structured = resp.get("structuredContent")
+        if isinstance(structured, dict) and isinstance(structured.get("result"), list):
+            return structured["result"]
+        content = resp.get("content")
+        if isinstance(content, list):
+            for part in content:
+                if isinstance(part, dict) and part.get("type") == "text":
+                    try:
+                        parsed = json.loads(part.get("text", "[]"))
+                    except json.JSONDecodeError:
+                        continue
+                    if isinstance(parsed, list):
+                        return parsed
+        if isinstance(content, str):
+            try:
+                parsed = json.loads(content)
+            except json.JSONDecodeError:
+                parsed = None
+            if isinstance(parsed, list):
+                return parsed
+
+    if isinstance(resp, list):
+        return resp
+
+    return []
 
 
 class AgentState(str, Enum):
@@ -46,7 +107,6 @@ class AgentTarget(str, Enum):
 
     LOCAL = "local"  # Issues requiring local machine access
     REMOTE = "remote"  # Issues that can run on cloud VM
-    ANY = "any"  # Process any issue regardless of target
 
 
 @dataclass
@@ -60,6 +120,7 @@ class AgentSlot:
     started_at: datetime | None = None
     completed_at: datetime | None = None
     error: str | None = None
+    work_key: str | None = None
 
 
 @dataclass
@@ -80,14 +141,14 @@ class AgentPool:
     def __init__(
         self,
         max_agents: int = MAX_CONCURRENT_AGENTS,
-        target: AgentTarget = AgentTarget.ANY,
+        target: AgentTarget = AgentTarget.REMOTE,
         max_issues_per_run: int = 0,
     ):
         """Initialize the agent pool.
 
         Args:
             max_agents: Maximum number of concurrent agents (default: 5)
-            target: Which issues to process (local, remote, or any)
+            target: Which issues to process (local or remote)
             max_issues_per_run: Limit issues processed per run (0 = unlimited)
         """
         self.max_agents = max_agents
@@ -97,15 +158,54 @@ class AgentPool:
         self._api_client: GitHubAPIClient | None = None
         self._projects_client: ProjectsV2Client | None = None
         self._issue_queue: IssueQueue | None = None
+        self._manager_agent: ManagerAgent | None = None
+        self._project_id: str | None = None
         self._running = False
         self._draining = False  # Drain mode: process until queue empty
         self._completed_count = 0
         self._failed_count = 0
-        self._processed_issues: set[int] = set()
+        self._fatal_error: str | None = None
+        self._processed_issues: set[str] = set()
         self._session_processed: int = 0  # Issues processed in current session
+        self._work_meta_by_key: dict[str, dict[str, Any]] = {}
         self._resume_completed: bool = False
         self._last_cleanup_at: datetime | None = None
+        self._refill_lock = asyncio.Lock()
+        self._refill_scheduled = False
         self.max_issues_per_run = max_issues_per_run
+
+    async def _refill_slots(self) -> None:
+        if self._fatal_error or not (self._running or self._draining):
+            return
+        async with self._refill_lock:
+            if self._fatal_error or not (self._running or self._draining):
+                return
+            await self.process_work_queue()
+
+    def _schedule_refill(self) -> None:
+        if self._refill_scheduled:
+            return
+        self._refill_scheduled = True
+
+        async def _run() -> None:
+            try:
+                await self._refill_slots()
+            finally:
+                self._refill_scheduled = False
+
+        asyncio.create_task(_run())
+
+    def _format_fatal_error(self, error: str) -> str:
+        message = (error or "unexpected failure").strip()
+        if not message.startswith("❌ ERROR"):
+            message = f"❌ ERROR: {message}"
+        return message
+
+    def _set_fatal_error(self, error: str) -> None:
+        if self._fatal_error:
+            return
+        self._fatal_error = self._format_fatal_error(error)
+        self.stop()
 
     @property
     def api_client(self) -> GitHubAPIClient:
@@ -134,6 +234,39 @@ class AgentPool:
             )
         return self._issue_queue
 
+    def _get_manager_agent(self) -> ManagerAgent | None:
+        if not self.settings.manager_agent_enabled:
+            return None
+        if self._manager_agent is None:
+            self._manager_agent = ManagerAgent()
+        return self._manager_agent
+
+    async def _hydrate_issue(self, issue: Issue) -> Issue:
+        if not issue.repo_owner or not issue.repo_name:
+            return issue
+        try:
+            full = await self.issue_queue.get_issue(
+                issue.number,
+                repo_owner=issue.repo_owner,
+                repo_name=issue.repo_name,
+            )
+            full.repo_owner = issue.repo_owner
+            full.repo_name = issue.repo_name
+            return full
+        except Exception as exc:
+            logger.warning(
+                "issue_hydration_failed",
+                issue=issue.number,
+                error=str(exc),
+            )
+            return issue
+
+    async def _hydrate_issues(self, issues: list[Issue]) -> list[Issue]:
+        hydrated: list[Issue] = []
+        for issue in issues:
+            hydrated.append(await self._hydrate_issue(issue))
+        return hydrated
+
     def get_status(self) -> PoolStatus:
         """Get current pool status."""
         active = sum(1 for s in self.slots if s.state == AgentState.RUNNING)
@@ -157,10 +290,97 @@ class AgentPool:
                 return slot
         return None
 
+    async def _get_project_id(self) -> str:
+        if self._project_id:
+            return self._project_id
+        project_id = await self.projects_client.get_org_project_id(
+            self.settings.github_org,
+            self.settings.github_project_name,
+        )
+        if not project_id:
+            raise ValueError(
+                f"Project '{self.settings.github_project_name}' not found in org "
+                f"'{self.settings.github_org}'"
+            )
+        self._project_id = project_id
+        return project_id
+
+    async def _fetch_blockers_via_appforge_mcp(self, issue: Issue) -> list[Issue]:
+        if not issue.repo_owner or not issue.repo_name:
+            return []
+
+        url = self.settings.appforge_mcp_url.rstrip("/")
+        if not url.endswith("/mcp"):
+            url = f"{url}/mcp"
+
+        try:
+            async with McpClient(url) as client:
+                resp = await client.call_tool(
+                    "list_issue_blockers",
+                    {
+                        "repo_owner": issue.repo_owner,
+                        "repo_name": issue.repo_name,
+                        "issue_number": issue.number,
+                    },
+                )
+        except Exception as exc:
+            raise ValueError(f"❌ ERROR: Failed to fetch issue blockers: {exc}") from exc
+
+        blockers: list[Issue] = []
+        now = datetime.now(UTC)
+        for item in _extract_mcp_items(resp):
+            try:
+                blockers.append(
+                    Issue(
+                        number=int(item["number"]),
+                        title=item.get("title", ""),
+                        body="",
+                        labels=item.get("labels", []),
+                        assignee=None,
+                        state=item.get("state", "open").lower(),
+                        created_at=now,
+                        updated_at=now,
+                        html_url=item.get("html_url", ""),
+                        repo_owner=item.get("repo_owner"),
+                        repo_name=item.get("repo_name"),
+                    )
+                )
+            except Exception as exc:  # pragma: no cover - defensive
+                logger.warning("mcp_blocker_parse_failed", item=item, error=str(exc))
+        return blockers
+
+    async def _has_blockers_not_done(self, issue: Issue) -> bool:
+        if not issue.repo_owner or not issue.repo_name:
+            return False
+        blockers = await self._fetch_blockers_via_appforge_mcp(issue)
+        if not blockers:
+            return False
+        project_id = await self._get_project_id()
+        not_done = []
+        for blocker in blockers:
+            status = await self.projects_client.get_issue_project_status(
+                project_id,
+                blocker.number,
+                blocker.repo_owner,
+                blocker.repo_name,
+            )
+            if status != IssueStatus.DONE.value:
+                not_done.append((blocker.number, status))
+        if not_done:
+            logger.debug(
+                "issue_skipped_blockers_not_done",
+                issue=issue.number,
+                blockers=not_done,
+            )
+            return True
+        return False
+
     def set_max_issues_per_run(self, limit: int) -> None:
         """Set the maximum issues to process in this run (0 = unlimited)."""
         self.max_issues_per_run = max(0, limit)
-        logger.info("max_issues_per_run_set", limit=self.max_issues_per_run, target=self.target.value)
+        logger.info(
+            "max_issues_per_run_set", limit=self.max_issues_per_run, target=self.target.value
+        )
 
     def _matches_target(self, issue: Issue) -> bool:
         """Check if issue matches the pool's target environment.
@@ -171,9 +391,6 @@ class AgentPool:
         Returns:
             True if issue should be processed by this pool
         """
-        if self.target == AgentTarget.ANY:
-            return True
-
         local_label = self.settings.github_local_agent_label
         remote_label = self.settings.github_remote_agent_label
 
@@ -181,13 +398,68 @@ class AgentPool:
         has_remote = remote_label in issue.labels
 
         if self.target == AgentTarget.LOCAL:
-            # Process only if explicitly marked local
-            return has_local
+            # Local can process either local or remote labeled work.
+            return has_local or has_remote
         elif self.target == AgentTarget.REMOTE:
             # Process only if explicitly marked remote
             return has_remote
 
         return True
+
+    def _issue_key(self, issue: Issue) -> str:
+        return f"issue:{issue.repo_owner}/{issue.repo_name}#{issue.number}"
+
+    def _filter_actionable(self, issues: list[Issue]) -> list[Issue]:
+        """Filter out issues that are blocked or waiting on developer input."""
+        labels = {
+            self.settings.github_remote_agent_label,
+            self.settings.github_local_agent_label,
+        }
+        return [issue for issue in issues if any(label in issue.labels for label in labels)]
+
+    async def _build_work_queue(self) -> tuple[list[tuple[Issue, str]], dict[str, int]]:
+        """Build an ordered work queue: in-progress first, then ready."""
+        in_progress = self._filter_actionable(await self.fetch_in_progress_issues())
+        ready = self._filter_actionable(await self.fetch_ready_issues())
+
+        counts = {
+            "in_progress": len(in_progress),
+            "ready": len(ready),
+        }
+
+        ordered: list[tuple[Issue, str]] = []
+        self._work_meta_by_key = {}
+        for issue in in_progress:
+            key = self._issue_key(issue)
+            ordered.append((issue, key))
+        for issue in ready:
+            key = self._issue_key(issue)
+            ordered.append((issue, key))
+
+        manager = self._get_manager_agent()
+        if not manager:
+            return ordered, counts
+
+        items = []
+        for issue in in_progress:
+            items.append({"category": "in_progress", "issue": issue, "key": self._issue_key(issue)})
+        for issue in ready:
+            items.append({"category": "ready", "issue": issue, "key": self._issue_key(issue)})
+
+        ordered_keys = await manager.order_work_items(items)
+        if not ordered_keys:
+            return ordered, counts
+
+        by_key = {key: issue for issue, key in ordered}
+        queue = []
+        for key in ordered_keys:
+            issue = by_key.get(key)
+            if issue:
+                queue.append((issue, key))
+        for issue, key in ordered:
+            if key not in ordered_keys:
+                queue.append((issue, key))
+        return queue, counts
 
     async def fetch_ready_issues(self) -> list[Issue]:
         """Fetch issues that are ready for processing.
@@ -200,44 +472,46 @@ class AgentPool:
         Returns:
             List of issues with "Ready" status and no open blockers
         """
-        # Prefer appforge MCP server when enabled and target is remote/any (server filters remote label + blockers)
-        if self.settings.appforge_mcp_enabled and self.target in {AgentTarget.REMOTE, AgentTarget.ANY}:
+        # Prefer appforge MCP server when enabled and target is remote (server filters remote label + blockers)
+        if self.settings.appforge_mcp_enabled and self.target == AgentTarget.REMOTE:
             mcp_issues = await self._fetch_ready_issues_via_mcp()
             if mcp_issues:
-                return mcp_issues
+                issues = mcp_issues
+            else:
+                issues = []
+        else:
+            issues = []
 
         try:
-            issues = await self.issue_queue.list_issues_by_project_status(
-                self.settings.github_project_name,
-                IssueStatus.READY.value,
-            )
+            if not issues:
+                issues = await self.issue_queue.list_issues_by_project_status(
+                    self.settings.github_project_name,
+                    IssueStatus.READY.value,
+                )
 
             # Filter out already processed issues
-            new_issues = [issue for issue in issues if issue.number not in self._processed_issues]
+            new_issues = [
+                issue for issue in issues if self._issue_key(issue) not in self._processed_issues
+            ]
 
             # Filter by target environment
             target_issues = [issue for issue in new_issues if self._matches_target(issue)]
 
-            # Filter out issues with open blockers
+            # Filter out issues with blockers not in Done status
             unblocked_issues = []
             blocked_count = 0
 
             for issue in target_issues:
-                if issue.repo_owner and issue.repo_name:
-                    has_blockers = await self.projects_client.has_open_blockers(
-                        issue.repo_owner,
-                        issue.repo_name,
-                        issue.number,
-                    )
-                    if has_blockers:
-                        blocked_count += 1
-                        logger.debug(
-                            "issue_skipped_has_blockers",
-                            issue=issue.number,
-                            title=issue.title,
-                        )
-                        continue
+                if await self._has_blockers_not_done(issue):
+                    blocked_count += 1
+                    continue
                 unblocked_issues.append(issue)
+
+            unblocked_issues = await self._hydrate_issues(unblocked_issues)
+            manager = self._get_manager_agent()
+            if manager:
+                selected = await manager.select_ready_issues(unblocked_issues)
+                unblocked_issues = [issue for issue in unblocked_issues if issue.number in selected]
 
             logger.info(
                 "fetched_ready_issues",
@@ -252,8 +526,9 @@ class AgentPool:
             return unblocked_issues
 
         except Exception as e:
-            logger.error("fetch_ready_issues_failed", error=str(e))
-            return []
+            error_message = f"❌ ERROR: fetch_ready_issues_failed: {e}"
+            logger.error("fetch_ready_issues_failed", error=error_message)
+            raise ValueError(error_message) from e
 
     async def _fetch_ready_issues_via_mcp(self) -> list[Issue]:
         """Fetch ready issues via appforge MCP server (already filtered by status/label/blockers)."""
@@ -270,8 +545,8 @@ class AgentPool:
                 }
                 resp = await client.call_tool("list_ready_remote_items", args)
                 issues: list[Issue] = []
-                now = datetime.utcnow()
-                for item in resp or []:
+                now = datetime.now(UTC)
+                for item in _extract_mcp_items(resp):
                     try:
                         issues.append(
                             Issue(
@@ -308,15 +583,30 @@ class AgentPool:
                 IssueStatus.IN_PROGRESS.value,
             )
 
+            issues = await self._hydrate_issues(issues)
             filtered = []
             for issue in issues:
-                if issue.number in self._processed_issues:
+                if self._issue_key(issue) in self._processed_issues:
                     continue
                 if not self._matches_target(issue):
                     continue
-                if self.settings.github_agent_label not in issue.labels:
+                if not self._filter_actionable([issue]):
+                    continue
+                if await self._has_blockers_not_done(issue):
+                    continue
+                if issue.assignee:
+                    logger.debug(
+                        "issue_skipped_assigned",
+                        issue=issue.number,
+                        assignee=issue.assignee,
+                    )
                     continue
                 filtered.append(issue)
+
+            manager = self._get_manager_agent()
+            if manager:
+                selected = await manager.select_resume_issues(filtered)
+                filtered = [issue for issue in filtered if issue.number in selected]
 
             logger.info(
                 "fetched_in_progress_issues",
@@ -325,8 +615,9 @@ class AgentPool:
             )
             return filtered
         except Exception as e:
-            logger.error("fetch_in_progress_issues_failed", error=str(e))
-            return []
+            error_message = f"❌ ERROR: fetch_in_progress_issues_failed: {e}"
+            logger.error("fetch_in_progress_issues_failed", error=error_message)
+            raise ValueError(error_message) from e
 
     async def resume_in_progress_issues(self) -> dict[str, Any]:
         """Resume issues that were in progress (startup sweep)."""
@@ -344,7 +635,7 @@ class AgentPool:
             if self.get_status().idle_slots == 0:
                 skipped += len(in_progress) - spawned - skipped
                 break
-            if await self.spawn_agent(issue):
+            if await self.spawn_agent(issue, self._issue_key(issue)):
                 spawned += 1
             else:
                 skipped += 1
@@ -384,6 +675,8 @@ class AgentPool:
 
         try:
             # Create initial state for the workflow
+            work_key = slot.work_key or self._issue_key(issue)
+            work_meta = self._work_meta_by_key.get(work_key, {})
             initial_state = WorkerState(
                 issue=issue,
                 issue_number=issue.number,
@@ -394,17 +687,13 @@ class AgentPool:
                     "repo_owner": issue.repo_owner,
                     "repo_name": issue.repo_name,
                     "repo": issue.repo_name,
+                    **work_meta,
                 },
             )
 
             # Run the workflow graph
             graph = get_compiled_graph()
             final_state = await graph.ainvoke(initial_state)
-
-            slot.state = AgentState.COMPLETED
-            slot.completed_at = datetime.now()
-            self._completed_count += 1
-            self._session_processed += 1
 
             # Handle both dict and WorkerState return types from graph
             agent_result = None
@@ -418,6 +707,7 @@ class AgentPool:
                 agent_result = final_state.agent_result
 
             result_status = "success"
+            status_value = None
             if agent_result:
                 if isinstance(agent_result, dict):
                     status_value = agent_result.get("status")
@@ -425,6 +715,21 @@ class AgentPool:
                     status_value = agent_result.status.value
                 if status_value != "success":
                     result_status = "failed"
+
+            if result_status == "failed":
+                error_message = None
+                if agent_result:
+                    if isinstance(agent_result, dict):
+                        error_message = agent_result.get("error") or agent_result.get("output")
+                    else:
+                        error_message = agent_result.error or agent_result.output
+                error_message = error_message or "agent reported failure"
+                raise RuntimeError(self._format_fatal_error(error_message))
+
+            slot.state = AgentState.COMPLETED
+            slot.completed_at = datetime.now()
+            self._completed_count += 1
+            self._session_processed += 1
 
             duration_seconds = (slot.completed_at - slot.started_at).total_seconds()
             metrics.observe_summary(
@@ -450,6 +755,7 @@ class AgentPool:
             slot.completed_at = datetime.now()
             slot.error = str(e)
             self._failed_count += 1
+            self._session_processed += 1
             duration_seconds = (slot.completed_at - slot.started_at).total_seconds()
             metrics.observe_summary(
                 "ace_agent_duration_seconds",
@@ -467,15 +773,18 @@ class AgentPool:
                 issue=issue.number,
                 error=str(e),
             )
+            self._set_fatal_error(str(e))
 
         finally:
             # Mark slot as idle for next issue
             slot.state = AgentState.IDLE
             slot.issue = None
             slot.task = None
+            slot.work_key = None
             metrics.dec_gauge("ace_active_agents", 1)
+            self._schedule_refill()
 
-    async def spawn_agent(self, issue: Issue) -> bool:
+    async def spawn_agent(self, issue: Issue, work_key: str) -> bool:
         """Spawn an agent for an issue if a slot is available.
 
         Args:
@@ -489,20 +798,27 @@ class AgentPool:
             logger.warning("no_idle_slots_available", issue=issue.number)
             return False
 
+        # Reserve the slot immediately to avoid oversubscribing slots.
+        slot.state = AgentState.RUNNING
+        slot.issue = issue
+
         # Mark as processed to avoid duplicate spawning
-        self._processed_issues.add(issue.number)
+        self._processed_issues.add(work_key)
+        slot.work_key = work_key
 
         # Create and store the task
         slot.task = asyncio.create_task(self._run_agent_for_issue(slot, issue))
         return True
 
-    async def process_ready_issues(self) -> dict[str, Any]:
-        """Fetch ready issues and spawn agents for them.
+    async def process_work_queue(self) -> dict[str, Any]:
+        """Fetch actionable issues and spawn agents for them.
 
         Returns:
             Summary of processing results
         """
-        logger.info("process_ready_issues_starting")
+        logger.info("process_work_queue_starting")
+        if self._fatal_error:
+            raise RuntimeError(self._fatal_error)
 
         limit = self.max_issues_per_run
         if limit > 0:
@@ -516,7 +832,129 @@ class AgentPool:
                     "pool_status": self.get_status().__dict__,
                 }
 
-        # Fetch ready issues from project board
+        work_queue, counts = await self._build_work_queue()
+
+        if not work_queue:
+            logger.info("no_actionable_issues_found", counts=counts)
+            return {
+                "status": "no_issues",
+                "spawned": 0,
+                "skipped": 0,
+                "pool_status": self.get_status().__dict__,
+            }
+
+        if limit > 0:
+            remaining = limit - self._session_processed
+            work_queue = work_queue[:remaining]
+
+        spawned = 0
+        skipped = 0
+
+        for issue, work_key in work_queue:
+            if self.get_status().idle_slots == 0:
+                logger.info("all_slots_busy", remaining_issues=len(work_queue) - spawned - skipped)
+                skipped += len(work_queue) - spawned - skipped
+                break
+
+            if await self.spawn_agent(issue, work_key):
+                spawned += 1
+            else:
+                skipped += 1
+
+        logger.info(
+            "process_work_queue_complete",
+            spawned=spawned,
+            skipped=skipped,
+            counts=counts,
+            pool_status=self.get_status().__dict__,
+        )
+
+        return {
+            "status": "processing",
+            "spawned": spawned,
+            "skipped": skipped,
+            "pool_status": self.get_status().__dict__,
+        }
+
+    async def process_in_progress_issues(self) -> dict[str, Any]:
+        """Fetch in-progress issues and spawn agents for them."""
+        logger.info("process_in_progress_issues_starting")
+        if self._fatal_error:
+            raise RuntimeError(self._fatal_error)
+
+        limit = self.max_issues_per_run
+        if limit > 0:
+            remaining = limit - self._session_processed
+            if remaining <= 0:
+                logger.info("max_issues_reached", max_issues=limit)
+                return {
+                    "status": "max_reached",
+                    "spawned": 0,
+                    "skipped": 0,
+                    "pool_status": self.get_status().__dict__,
+                }
+
+        issues = await self.fetch_in_progress_issues()
+        if not issues:
+            logger.info("no_in_progress_issues_found")
+            return {
+                "status": "no_issues",
+                "spawned": 0,
+                "skipped": 0,
+                "pool_status": self.get_status().__dict__,
+            }
+
+        if limit > 0:
+            remaining = limit - self._session_processed
+            issues = issues[:remaining]
+
+        spawned = 0
+        skipped = 0
+        for issue in issues:
+            if self.get_status().idle_slots == 0:
+                logger.info("all_slots_busy", remaining_issues=len(issues) - spawned - skipped)
+                skipped += len(issues) - spawned - skipped
+                break
+            if await self.spawn_agent(issue, self._issue_key(issue)):
+                spawned += 1
+            else:
+                skipped += 1
+
+        logger.info(
+            "process_in_progress_issues_complete",
+            spawned=spawned,
+            skipped=skipped,
+            pool_status=self.get_status().__dict__,
+        )
+        return {
+            "status": "processing",
+            "spawned": spawned,
+            "skipped": skipped,
+            "pool_status": self.get_status().__dict__,
+        }
+
+    async def process_ready_issues(self) -> dict[str, Any]:
+        """Fetch ready issues and spawn agents for them.
+
+        Returns:
+            Summary of processing results
+        """
+        logger.info("process_ready_issues_starting")
+        if self._fatal_error:
+            raise RuntimeError(self._fatal_error)
+
+        limit = self.max_issues_per_run
+        if limit > 0:
+            remaining = limit - self._session_processed
+            if remaining <= 0:
+                logger.info("max_issues_reached", max_issues=limit)
+                return {
+                    "status": "max_reached",
+                    "spawned": 0,
+                    "skipped": 0,
+                    "pool_status": self.get_status().__dict__,
+                }
+
         ready_issues = await self.fetch_ready_issues()
 
         if not ready_issues:
@@ -543,7 +981,7 @@ class AgentPool:
                 skipped += len(ready_issues) - spawned - skipped
                 break
 
-            if await self.spawn_agent(issue):
+            if await self.spawn_agent(issue, self._issue_key(issue)):
                 spawned += 1
             else:
                 skipped += 1
@@ -572,22 +1010,22 @@ class AgentPool:
         self._session_processed = 0
         logger.info("agent_pool_starting", max_agents=self.max_agents, poll_interval=poll_interval)
 
-        if self.settings.resume_in_progress_issues:
-            try:
-                await self.resume_in_progress_issues()
-            except Exception as e:
-                logger.error("resume_in_progress_failed", error=str(e))
+        try:
+            while self._running:
+                if self._fatal_error:
+                    raise RuntimeError(self._fatal_error)
+                await self.process_work_queue()
 
-        while self._running:
-            try:
-                await self.process_ready_issues()
-            except Exception as e:
-                logger.error("pool_cycle_error", error=str(e))
+                await self._maybe_cleanup()
 
-            await self._maybe_cleanup()
+                if self._fatal_error:
+                    raise RuntimeError(self._fatal_error)
 
-            # Wait before next poll
-            await asyncio.sleep(poll_interval)
+                # Wait before next poll
+                await asyncio.sleep(poll_interval)
+        except Exception as e:
+            self._set_fatal_error(str(e))
+            raise RuntimeError(self._fatal_error) from e
 
         logger.info("agent_pool_stopped")
 
@@ -610,31 +1048,48 @@ class AgentPool:
 
         logger.info("drain_mode_starting", max_agents=self.max_agents)
 
-        while self._draining:
-            # Fetch and spawn for any ready issues
-            result = await self.process_ready_issues()
+        try:
+            while self._draining:
+                if self._fatal_error:
+                    raise RuntimeError(self._fatal_error)
+                # Fetch and spawn for actionable work
+                result = await self.process_work_queue()
+                if self._fatal_error:
+                    raise RuntimeError(self._fatal_error)
 
-            status = self.get_status()
+                status = self.get_status()
 
-            # Check if we're done:
-            # - No issues were spawned or available
-            # - All agents are idle (no active work)
-            if result["spawned"] == 0 and status.active_agents == 0:
-                # Double-check by fetching again (in case blockers just resolved)
-                ready_issues = await self.fetch_ready_issues()
-                if not ready_issues:
+                if result.get("status") == "max_reached":
                     logger.info(
-                        "drain_mode_complete",
+                        "drain_mode_max_issues_reached",
                         session_processed=self._session_processed,
-                        duration_seconds=(datetime.now() - session_start).total_seconds(),
                     )
                     break
 
-            # Wait before next check
-            await self._maybe_cleanup()
-            await asyncio.sleep(check_interval)
+                # Check if we're done:
+                # - No issues were spawned or available
+                # - All agents are idle (no active work)
+                if result["spawned"] == 0 and status.active_agents == 0:
+                    # Double-check by fetching again (in case blockers just resolved)
+                    work_queue, _ = await self._build_work_queue()
+                    if not work_queue:
+                        logger.info(
+                            "drain_mode_complete",
+                            session_processed=self._session_processed,
+                            duration_seconds=(datetime.now() - session_start).total_seconds(),
+                        )
+                        break
 
-        self._draining = False
+                # Wait before next check
+                await self._maybe_cleanup()
+                if self._fatal_error:
+                    raise RuntimeError(self._fatal_error)
+                await asyncio.sleep(check_interval)
+        except Exception as e:
+            self._set_fatal_error(str(e))
+            raise RuntimeError(self._fatal_error) from e
+        finally:
+            self._draining = False
 
         return {
             "status": "complete",
@@ -739,6 +1194,11 @@ class AgentPool:
                 "cleanup_tmux_session",
                 session=session_name,
             )
+            log_key_event(
+                logger,
+                "🧹 tmux session cleaned up",
+                session=session_name,
+            )
             tmux.kill_session(session_name)
 
     async def wait_for_completion(self, timeout: float | None = None) -> None:
@@ -765,11 +1225,11 @@ class AgentPool:
 _pools: dict[AgentTarget, AgentPool] = {}
 
 
-def get_pool(target: AgentTarget = AgentTarget.ANY) -> AgentPool:
+def get_pool(target: AgentTarget = AgentTarget.REMOTE) -> AgentPool:
     """Get or create an agent pool for the specified target.
 
     Args:
-        target: Which issues to process (local, remote, or any)
+        target: Which issues to process (local or remote)
 
     Returns:
         AgentPool instance for the specified target
