@@ -1,3 +1,3 @@
-"""Agentic Coding Engine - orchestrates agents to handle GitHub issues."""
+"""Appforge Coding Engine - orchestrates agents to handle GitHub issues."""
 
 __version__ = "0.1.0"
